# app/v1/swagger.yml
...
swagger: '2.0'
...
info:
  version: '1.0.0'
  title: N0deJS API
  description: N0deJS REST API prototype
schemes:
  - http
host: localhost:3001
basePath: /v1
...
paths:
...
  '/projects/{projectId}/':

    get:
      summary: Get the project specified by `projectId`.
      description: |
        Get the project specified in the request. If the
        item is not found, return a `404`.
      operationId: getProject
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed.
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: projectId
          in: path
          description: The id of the project to get.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response with a `project` object
          schema:
            $ref: '#/definitions/project'
        # Response code
        404:
          description: The specified project is not found.

    put:
      summary: Update the project specified by `projectId`.
      description: |
        Update the specified project and return the updated project object.
      operationId: updateProject
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectId
          in: path
          description: The id of the project.
          required: true
          type: integer
          format: int32

        - name: project
          required: true
          in: body
          description: The project object.
          schema:
            $ref: '#/definitions/project'

      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: The updated project.
          schema:
            $ref: '#/definitions/project'
        # Response code
        default:
          description: Failed to update the project.
          schema:
            $ref: '#/definitions/error'

    delete:
      summary: Delete the project specified by `projectId`.
      description: |
        Delete the project referenced in the request.
        If the project is not found, return `404`.
      operationId: deleteProject
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed.
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: projectId
          in: path
          description: The id of the project to delete.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
        # Response code
        204:
          description: Successfully deleted.
        # Response code
        404:
          description: When environment with specified id is not found or there are projects using it.
        default:
          description: Unexpected error.
...
definitions:
...
  branch:
    type: object
    description: |
      A branch of a repo.
    required:
      - name
    properties:
      name:
        type: string
        description: The branch name, e.g. `master`.
      allCommitsLoaded:
        type: boolean
        description: |
          This is `true` if all commits for this branch have been loaded. TBD: This looks like some UI-related funkiness.
      commitPagesLoaded:
        type: integer
        format: int32
        description: The number of commit pages loaded (for the purposes of paging through the set of commits).
      commit:
        $ref: '#/definitions/commit'
      commits:
        type: array
        description: An array of `commit` objects, which can be empty.
        items:
          $ref: '#/definitions/commit'


  ############## Model: commit ##############
  commit:
    type: object
    description: |
      A commit to a repo.
    required:
      - sha
      - url
    properties:
      sha:
        type: string
        description: The SHA of the commit, e.g. `16db3f928436de3cb14179f6a2d6e86220d6943c`.
      url:
        type: string
        description: The URL of the commit, e.g. `https://github.com/foobar/foobar.github.io/commit/16db3f928436de3cb14179f6a2d6e86220d6943c`.


  ############## Model: deployment ##############
...
  error:
    type: object
    description: |
      A standard error object returned by all API calls. TBD: this should be flattened into a single object,
      rather than having a nested `error` object.
    required:
      - apiVersion
      - error
    properties:
      apiVersion:
        type: string
        description: The API version, e.g. `v1`.
      error:
        type: object
        description: A standard error object returned by all API calls.
        required:
          - code
          - message
        properties:
          code:
            type: string
            description: The error code, usually an integer, often a HTTP status code.
          message:
            type: string
            description: The message associated with the error.
          log:
            type: string
            description: Any additional error information, such as log entries, stack trace, etc. Likely to be turned off in production environments.


  ############## Model: member ##############
  member:
    type: object
    description: |
      A member (user) of a project.
    required:
      - uid
    properties:
      uid:
        type: integer
        format: int32
        description: The user id of the project member.
      projectId:
        type: integer
        format: int32
        description: The id of the project that the user is a member of.


  ############## Model: members ##############
...
  project:
    type: object
    description: |
      Models an individual project.
    required:
      - id
      - name
      - type
    properties:
      id:
        type: integer
        format: int32
        description: The internal project id.
      name:
        type: string
        description: The user-provided label for the project. # This field should be renamed to 'label'
      'type':
        type: string
        description: The platform type of the project, e.g. 'nodejs' or 'java', etc.
      lastCommitSha:
        type: string
        description: The last git commit SHA.
      targets:
        type: array
        description: An array of target deployment environments (such as a CloudFoundry or AWS instance) that the project is deployed to.
        items:
          $ref: '#/definitions/target'
      userId:
        type: integer
        format: int32
        description: The internal user id of the project owner.
      joinCode:
        type: string
        description: The (semi-secret) code that an invited user users to join the project.
      repo:
        $ref: '#/definitions/repo'
      token:
        type: string
        description: The GitHub token to use (presumably???)
      branchRefName:
        type: string
        description: The branch name, e.g.`master`
      cloudFoundryEnvironment:
        # description: The (FIXME: HARDCODED!) Cloud Foundry environment that this project is deployed to.
        $ref: '#/definitions/target'
      members:
        type: array
        description: The members of this project.
        items:
          $ref: '#/definitions/member'


  ############## Model: projects ##############
...
  repo:
    type: object
    description: Models a Version Control System repository, such as a GitHub repo, and references a specific branch.
    required:
      - vcs
      - owner
      - name
      - branch
      - cloneUrl
      - sshUrl
      - httpUrl

    properties:
      vcs:
        type: string
        description: The repo type, e.g. `github`
      full_name:
        type: string
        description: The full name of the repo (typically `owner/repo`), e.g. `foobar/foobar.github.io`.
      owner:
        type: string
        description: The owner of the repo (as determined by the repo itself), e.g. `foobar`
      name:
        type: string
        description: The name of the repo, e.g. `hello-node`.
      githubRepoId:
        type: string
        description: For GitHub repos, the unique GitHub identifier.
      branch:
        type: string
        description: The repo branch being referenced, e.g. `master`.
      cloneUrl:
        type: string
        description: The HTTPS URL used to clone the repo.
      sshUrl:
        type: string
        description: The SSH URL used to clone the repo.
      httpUrl:
        type: string
        description: The repo home page.
      webHookId:
        type: string
        description: On GitHub, this is the unique id of the webhook for this repo that calls back to N0deJSAPI.
      secret:
        type: string
        description: The repo secret used to create a web-hook..
      branches:
        type: array
        description: |
          The set of branches for this repo.
        items:
          $ref: '#/definitions/branch'


  ############## Model: target ##############
  target:
    type: object
    description: Models a deployment target/environment (e.g. a CloudFoundry instance) that a project can be deployed to.
    required:
      - id
      - type
      - url
      - userName
      - password
      - name
      - organization
      - space
    properties:
      id:
        type: integer
        format: int32
        description: The internal id of this target instance.
      'type':
        type: string
        description: The type of target environment, e.g. `cloudfoundry`, `aws`, etc..
      url:
        type: string
        description: The URL endpoint that the target is accessible at.
      userName:
        type: string
        description: The username to authenticate to the target with.
      'password':
        type: string
        description: The password to authenticate to the target with.
      name:
        type: string
        description: The user-provided label for this target. # This field should be renamed to label
      organization:
        type: string
        description: The organization under which a project will be deployed to this target.
      space:
        type: string
        description: The space within an organization that a project will be deployed under on this target.


  ############## Model: targets ##############
...